refactor: adjust tests for activity bar rendering
