+++ packages/renderer-worker/test/ViewletActivityBar.test.js
@@ -115,124 +115,6 @@ test('loadContent', async () => {
   })
 })
 
-test('render - all items fit but little space is remaining', async () => {
-  const oldState = ViewletActivityBar.create()
-  const newState = {
-    ...oldState,
-    height: 369,
-    activityBarItems: [
-      // Top
-      {
-        id: 'Explorer',
-        title: 'Explorer',
-        icon: 'icons/files.svg',
-        enabled: true,
-        flags: ActivityBarItemFlags.Tab,
-        keyShortcuts: 'Control+Shift+E',
-      },
-      {
-        id: 'Search',
-        title: 'Search',
-        icon: 'icons/search.svg',
-        enabled: true,
-        flags: ActivityBarItemFlags.Tab,
-        keyShortcuts: 'Control+Shift+F',
-      },
-      {
-        id: 'Source Control',
-        title: 'Source Control',
-        icon: 'icons/source-control.svg',
-        enabled: true,
-        flags: ActivityBarItemFlags.Tab,
-        keyShortcuts: 'Control+Shift+G',
-      },
-      {
-        id: 'Run and Debug',
-        title: 'Run and Debug',
-        icon: 'icons/debug-alt-2.svg',
-        enabled: true,
-        flags: ActivityBarItemFlags.Tab,
-        keyShortcuts: 'Control+Shift+D',
-      },
-      {
-        id: 'Extensions',
-        title: 'Extensions',
-        icon: 'icons/extensions.svg',
-        enabled: true,
-        flags: ActivityBarItemFlags.Tab,
-        keyShortcuts: 'Control+Shift+X',
-      },
-      // Bottom
-      {
-        id: 'Settings',
-        title: 'Settings',
-        icon: 'icons/settings-gear.svg',
-        enabled: true,
-        flags: ActivityBarItemFlags.Button,
-        keyShortcuts: '',
-      },
-    ],
-  }
-
-  expect(render(oldState, newState)).toEqual([
-    [
-      'Viewlet.send',
-      'ActivityBar',
-      'setItems',
-      [
-        {
-          enabled: true,
-          flags: ActivityBarItemFlags.Tab,
-          icon: 'icons/files.svg',
-          id: 'Explorer',
-          title: 'Explorer',
-          keyShortcuts: 'Control+Shift+E',
-        },
-        {
-          enabled: true,
-          flags: ActivityBarItemFlags.Tab,
-          icon: 'icons/search.svg',
-          id: 'Search',
-          title: 'Search',
-          keyShortcuts: 'Control+Shift+F',
-        },
-        {
-          enabled: true,
-          flags: ActivityBarItemFlags.Tab,
-          icon: 'icons/source-control.svg',
-          id: 'Source Control',
-          title: 'Source Control',
-          keyShortcuts: 'Control+Shift+G',
-        },
-        {
-          enabled: true,
-          flags: ActivityBarItemFlags.Tab,
-          icon: 'icons/debug-alt-2.svg',
-          id: 'Run and Debug',
-          title: 'Run and Debug',
-          keyShortcuts: 'Control+Shift+D',
-        },
-        {
-          enabled: true,
-          flags: ActivityBarItemFlags.Tab,
-          icon: 'icons/extensions.svg',
-          id: 'Extensions',
-          title: 'Extensions',
-          keyShortcuts: 'Control+Shift+X',
-        },
-        {
-          enabled: true,
-          flags: ActivityBarItemFlags.Button,
-          icon: 'icons/settings-gear.svg',
-          id: 'Settings',
-          title: 'Settings',
-          keyShortcuts: '',
-        },
-      ],
-    ],
-  ])
-})
-
 test.skip('contentLoaded - one items does not fit', async () => {
   // @ts-ignore
   RendererProcess.invoke.mockImplementation(() => {})
@@ -342,116 +224,6 @@ test.skip('contentLoaded - one items does not fit', async () => {
       },
     ],
     0,
-  ])
-})
-
-test('render - two items do not fit', () => {
-  const oldState = ViewletActivityBar.create()
-  const newState = {
-    ...oldState,
-    height: oldState.itemHeight * 5,
-    activityBarItems: [
-      // Top
-      {
-        id: 'Explorer',
-        title: 'Explorer',
-        icon: 'icons/files.svg',
-        enabled: true,
-        flags: ActivityBarItemFlags.Tab,
-        keyShortcuts: 'Control+Shift+E',
-      },
-      {
-        id: 'Search',
-        title: 'Search',
-        icon: 'icons/search.svg',
-        enabled: true,
-        flags: ActivityBarItemFlags.Tab,
-        keyShortcuts: 'Control+Shift+F',
-      },
-      {
-        id: 'Source Control',
-        title: 'Source Control',
-        icon: 'icons/source-control.svg',
-        enabled: true,
-        flags: ActivityBarItemFlags.Tab,
-        keyShortcuts: 'Control+Shift+G',
-      },
-      {
-        id: 'Run and Debug',
-        title: 'Run and Debug',
-        icon: 'icons/debug-alt-2.svg',
-        enabled: true,
-        flags: ActivityBarItemFlags.Tab,
-        keyShortcuts: 'Control+Shift+D',
-      },
-      {
-        id: 'Extensions',
-        title: 'Extensions',
-        icon: 'icons/extensions.svg',
-        enabled: true,
-        flags: ActivityBarItemFlags.Tab,
-        keyShortcuts: 'Control+Shift+X',
-      },
-      // Bottom
-      {
-        id: 'Settings',
-        title: 'Settings',
-        icon: 'icons/settings-gear.svg',
-        enabled: true,
-        flags: ActivityBarItemFlags.Button,
-        keyShortcuts: '',
-      },
-    ],
-  }
-
-  expect(render(oldState, newState)).toEqual([
-    [
-      'Viewlet.send',
-      'ActivityBar',
-      'setItems',
-      [
-        {
-          enabled: true,
-          flags: ActivityBarItemFlags.Tab,
-          icon: 'icons/files.svg',
-          id: 'Explorer',
-          title: 'Explorer',
-          keyShortcuts: 'Control+Shift+E',
-        },
-        {
-          enabled: true,
-          flags: ActivityBarItemFlags.Tab,
-          icon: 'icons/search.svg',
-          id: 'Search',
-          title: 'Search',
-          keyShortcuts: 'Control+Shift+F',
-        },
-        {
-          enabled: true,
-          flags: ActivityBarItemFlags.Tab,
-          icon: 'icons/source-control.svg',
-          id: 'Source Control',
-          title: 'Source Control',
-          keyShortcuts: 'Control+Shift+G',
-        },
-        {
-          enabled: true,
-          flags: ActivityBarItemFlags.Button,
-          icon: 'icons/ellipsis.svg',
-          id: 'Additional Views',
-          title: 'Additional Views',
-          keyShortCuts: '',
-        },
-        {
-          enabled: true,
-          flags: ActivityBarItemFlags.Button,
-          icon: 'icons/settings-gear.svg',
-          id: 'Settings',
-          title: 'Settings',
-          keyShortcuts: '',
-        },
-      ],
-    ],
   ])
 })
